+++ apistar/codecs/__init__.py
@@ -1,13 +1,16 @@
 from apistar.codecs.base import BaseCodec
+from apistar.codecs.config import ConfigCodec
 from apistar.codecs.download import DownloadCodec
 from apistar.codecs.jsondata import JSONCodec
 from apistar.codecs.jsonschema import JSONSchemaCodec
 from apistar.codecs.multipart import MultiPartCodec
 from apistar.codecs.openapi import OpenAPICodec
+from apistar.codecs.swagger import SwaggerCodec
 from apistar.codecs.text import TextCodec
 from apistar.codecs.urlencoded import URLEncodedCodec
 
 __all__ = [
-    'BaseCodec', 'JSONCodec', 'JSONSchemaCodec', 'OpenAPICodec', 'TextCodec',
-    'DownloadCodec', 'MultiPartCodec', 'URLEncodedCodec',
+    'BaseCodec', 'ConfigCodec', 'JSONCodec', 'JSONSchemaCodec', 'OpenAPICodec',
+    'SwaggerCodec', 'TextCodec', 'DownloadCodec', 'MultiPartCodec',
+    'URLEncodedCodec',
 ]